requirements.txt
flask
flask-cors
flask-sqlalchemy
psycopg2-binary
openai
streamlit==1.24.0
fastapi==0.100.0
uvicorn==0.23.1
sqlalchemy==2.0.19
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.1.0
jinja2==3.1.2
python-dotenv==1.0.0
pydantic[email]==2.1.1
pandas==2.0.3
requests==2.31.0
PyJWT==2.8.0
